Rename AnnoDesigner.model -> AnnoDesigner.Models
