+++ AnnoDesigner/App.xaml.cs
@@ -6,7 +6,7 @@ using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
-using AnnoDesigner.model;
+using AnnoDesigner.Models;
 using AnnoDesigner.viewmodel;
 using NLog;
 using NLog.Targets;
