+++ AnnoDesigner/CustomEventArgs/UpdateStatisticsEventArgs.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using AnnoDesigner.model;
+using AnnoDesigner.Models;
 
 namespace AnnoDesigner.CustomEventArgs
 {
