+++ AnnoDesigner/Models/ICanUpdateLayout.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AnnoDesigner.model
+namespace AnnoDesigner.Models
 {
     public interface ICanUpdateLayout
     {
