+++ AnnoDesigner/Models/UpdateMode.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AnnoDesigner.model
+namespace AnnoDesigner.Models
 {
     public enum UpdateMode
     {
