+++ Tests/AnnoDesigner.Tests/PresetsTreeViewModelTests.cs
@@ -7,8 +7,8 @@ using AnnoDesigner.Core;
 using AnnoDesigner.Core.Models;
 using AnnoDesigner.Core.Presets.Loader;
 using AnnoDesigner.Core.Presets.Models;
-using AnnoDesigner.model;
-using AnnoDesigner.model.PresetsTree;
+using AnnoDesigner.Models;
+using AnnoDesigner.Models.PresetsTree;
 using AnnoDesigner.viewmodel;
 using Moq;
 using Xunit;
